1.0.4
Repeated calls: Support for repeated reads or repeated writes. What this means is that it was not possible to read and read again, because the previous call was overriding the next call and breaking its data structure.
Start, restart: Now it is more reliable to restart the server.
Updated test cases: The test cases were broken due to the issue that was fixed with "Repeated calls" change.
Classes are separated from each other and most-used variables are now more properly named