Avoid consistency breaking logic
If the data was larger than available, we were only sending as much as available and sparing the error for the end.

Now, bound checking is forced upon the caller. In other words, if data may be larger than reported, the user should ensure the length is truncated before the invocation.