impl: Add the roaming state to ActiveConnection
In some of the use cases, I needed the active
connection to keep living after the server thread
that spawned it went out of scope. This change
makes it possible by adding a new state called
'roaming' which allows the server (or any other
owner) to skip closing the instance.